refactor ImpactWallet
