+++ app/src/ProofOfWork/ActionBar/BossMode/BossMode.styles.tsx
