+++ app/src/ProofOfWork/ActionBar/BossMode/BossMode.tsx
