+++ app/src/ProofOfWork/ActionBar/BossMode/Templates/MultipleChoiceQuestion/MultipleChoiceQuestion.tsx
