+++ app/src/ProofOfWork/ActionBar/BossMode/Templates/OutputQuestion/OutputQuestion.tsx
