+++ app/src/ProofOfWork/ActionBar/BossMode/Templates/SelectionQuestion/SelectionQuestion.tsx
