+++ app/src/ProofOfWork/ActionBar/BossMode/Templates/TextInputQuestion/TextInputQuestion.tsx
