+++ app/src/ProofOfWork/ActionBar/Cofounder/Cofounder.compute.tsx
