+++ app/src/ProofOfWork/ActionBar/Cofounder/Cofounder.constants.ts
