+++ app/src/ProofOfWork/ActionBar/Cofounder/Cofounder.tsx
