+++ app/src/ProofOfWork/ActionBar/EmotionalIntelligence/EmotionalIntelligence.compute.tsx
