+++ app/src/ProofOfWork/ActionBar/EmotionalIntelligence/EmotionalIntelligence.data.tsx
