+++ app/src/ProofOfWork/ActionBar/EmotionalIntelligence/SunsetCanvas.tsx
