--- app/src/ProofOfWork/ImpactWallet/ImpactWallet.styles.tsx
@@ -1,12 +1,0 @@
-import styled, { keyframes } from "styled-components";
-
-export const StarsContainer = styled.div`
-  position: relative;
-  /* Add additional styling as needed */
-`;
-
-export const Star = styled.div`
-  position: absolute;
-  opacity: 0; /* Initially hidden */
-  /* Basic styling for stars */
-`;
